Update verifiedCampaigns endpoint implemented
