+++ config/default.json
@@ -194,5 +194,9 @@
     "whitelist": []
   },
   "givethIoUrl": "https://serve.giveth.io/graphql",
-  "givethIoProjectsReviewerAddress": "0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1"
+  "givethIoProjectsReviewerAddress": "0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1",
+  "givethIoInfo": {
+    "username": "username",
+    "password": "password"
+  }
 }
