+++ src/services/index.js
@@ -18,7 +18,6 @@ const transactions = require('./transactions/transactions.service.js');
 const whitelist = require('./whitelist/whitelist.service.js');
 const gasprice = require('./gasprice/gasprice.service.js');
 const verifiedCampaigns = require('./verifiedCampaigns/verifiedCampaigns.service');
-const verifyCampaigns = require('./verifyCampaigns/verifyCampaigns.service');
 const conversionRates = require('./conversionRates/conversionRates.service.js');
 const conversations = require('./conversations/conversations.service.js');
 const givbackReportDonations = require('./verifiedProjectsGiversReport/verifiedPerojectsGiversReport.service');
@@ -47,5 +46,4 @@ module.exports = function configure() {
   app.configure(campaigncsv);
   app.configure(givbackReportDonations);
   app.configure(verifiedCampaigns);
-  app.configure(verifyCampaigns);
 };
