+++ src/services/verifiedCampaigns/verifiedCampaigns.hooks.js
@@ -2,6 +2,6 @@ const { authorizeGivethio } = require('../../hooks/authorizeGivethio');
 
 module.exports = {
   before: {
-    create: [authorizeGivethio()],
+    update: [authorizeGivethio()],
   },
 };
