+++ src/utils/swaggerUtils.test.js
@@ -5,10 +5,13 @@ const { generateRandomNumber } = require('../../test/testUtility');
 
 const generateSwaggerDocForCRUDServiceTestCases = () => {
   it('should return swagger schema from mongoose model', () => {
-    const testModel = mongoose.model(`test-${generateRandomNumber(1, 10000)}`, {
-      name: String,
-      lastName: String,
-    });
+    const testModel = mongoose.model(
+      `test-${new Date() + new Date().getMilliseconds() + generateRandomNumber(0, 1000)}`,
+      {
+        name: String,
+        lastName: String,
+      },
+    );
     const generatedSchema = generateSwaggerDocForCRUDService({ Model: testModel });
     assert.exists(generatedSchema.securities);
     assert.exists(generatedSchema.definition);
@@ -20,32 +23,41 @@ const generateSwaggerDocForCRUDServiceTestCases = () => {
     assert.equal(generatedSchema.operations.find.parameters[1].name, '$skip');
   });
   it('should disabledMethods set method false', () => {
-    const testModel = mongoose.model(`test-${generateRandomNumber(1, 10000)}`, {
-      name: String,
-      lastName: String,
-      age: Number,
-    });
+    const testModel = mongoose.model(
+      `test-${new Date() + new Date().getMilliseconds() + generateRandomNumber(0, 1000)}`,
+      {
+        name: String,
+        lastName: String,
+        age: Number,
+      },
+    );
     const generatedSchema = generateSwaggerDocForCRUDService({ Model: testModel }, ['create']);
     assert.equal(generatedSchema.operations.create, false);
     assert.notExists(generatedSchema.operations.remove);
   });
   it('should add fields to operation find parameters', () => {
-    const testModel = mongoose.model(`test-${generateRandomNumber(1, 10000)}`, {
-      name: String,
-      lastName: String,
-      age: Number,
-    });
+    const testModel = mongoose.model(
+      `test-${new Date() + new Date().getMilliseconds() + generateRandomNumber(0, 1000)}`,
+      {
+        name: String,
+        lastName: String,
+        age: Number,
+      },
+    );
     const generatedSchema = generateSwaggerDocForCRUDService({ Model: testModel });
     assert.equal(generatedSchema.operations.find.parameters[2].name, 'name');
     assert.equal(generatedSchema.operations.find.parameters[3].name, 'lastName');
   });
   it('should not add non-string fields to find operation parameters', () => {
-    const testModel = mongoose.model(`test-${generateRandomNumber(1, 10000)}`, {
-      name: String,
-      lastName: String,
-      age: Number,
-      alive: Boolean,
-    });
+    const testModel = mongoose.model(
+      `test-${new Date() + new Date().getMilliseconds() + generateRandomNumber(0, 1000)}`,
+      {
+        name: String,
+        lastName: String,
+        age: Number,
+        alive: Boolean,
+      },
+    );
     const generatedSchema = generateSwaggerDocForCRUDService({ Model: testModel });
     assert.equal(generatedSchema.operations.find.parameters[2].name, 'name');
     assert.equal(generatedSchema.operations.find.parameters[3].name, 'lastName');
@@ -53,12 +65,15 @@ const generateSwaggerDocForCRUDServiceTestCases = () => {
     assert.notExists(generatedSchema.operations.find.parameters[5]);
   });
   it('should not add verified field in find operation', () => {
-    const testModel = mongoose.model(`test-${generateRandomNumber(1, 10000)}`, {
-      name: String,
-      lastName: String,
-      age: Number,
-      alive: Boolean,
-    });
+    const testModel = mongoose.model(
+      `test-${new Date() + new Date().getMilliseconds() + generateRandomNumber(0, 1000)}`,
+      {
+        name: String,
+        lastName: String,
+        age: Number,
+        alive: Boolean,
+      },
+    );
     const generatedSchema = generateSwaggerDocForCRUDService({ Model: testModel });
     assert.equal(generatedSchema.operations.find.parameters[2].name, 'name');
     assert.equal(generatedSchema.operations.find.parameters[3].name, 'lastName');
@@ -66,11 +81,14 @@ const generateSwaggerDocForCRUDServiceTestCases = () => {
     assert.notExists(generatedSchema.operations.find.parameters[5]);
   });
   it('should add verified field in find operation', () => {
-    const testModel = mongoose.model(`test-${generateRandomNumber(1, 10000)}`, {
-      name: String,
-      lastName: String,
-      verified: Boolean,
-    });
+    const testModel = mongoose.model(
+      `test-${new Date() + new Date().getMilliseconds() + generateRandomNumber(0, 1000)}`,
+      {
+        name: String,
+        lastName: String,
+        verified: Boolean,
+      },
+    );
     const generatedSchema = generateSwaggerDocForCRUDService({ Model: testModel });
     assert.equal(generatedSchema.operations.find.parameters[2].name, 'name');
     assert.equal(generatedSchema.operations.find.parameters[3].name, 'lastName');
